requirements.txt
numpy>=1.20.0
pygame>=2.0.1
pymunk>=6.0.0
pyaudio>=0.2.11
librosa>=0.8.1
soundfile>=0.10.3
pydub>=0.25.1
pycryptodome>=3.10.1
tqdm>=4.61.0
matplotlib>=3.4.3
nltk>=3.6.3
pillow>=8.3.2
scikit-learn>=1.0.0
gensim>=4.1.2
flask>=2.0.1
PyOpenGL>=3.1.0
pyglet>=1.5.0
websockets>=10.0
tk>=0.1.0
requests>=2.25.1
scipy>=1.7.0 